requirements.txt
alembic==1.16.5
aniso8601==10.0.1
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cors==1.0.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.8.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.1
mysqlclient==2.2.7
passlib==1.7.4
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-file==2.1.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
tldextract==5.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==5.1.1
zope.interface==7.2
